backend/requirements.docker.txt
fastapi==0.136.0
uvicorn==0.46.0
python-multipart==0.0.26
python-dotenv==1.2.2
google-cloud-storage==3.10.1
opencv-python-headless==4.7.0.72
torch==2.5.1
torchvision==0.20.1
faiss-cpu==1.9.0
numpy==1.26.4
pydantic==2.13.3
